refactor: Move node utils to lib folder
We need to share implementationsa between the main and worker parts,
therefore we factor the common parts in a library directory. The IPC
parts could fit in there from a certain POV, but are also shared with
the renderer. For now I'd like to keep things which might end up in the
renderer outside of the library.