feat: Add on-demand job scheduling aka `JobEmitter`
We want to use the job system for artifact tasks. However, the current
job system runs on a strict schedule (pun intended). We need to be able
to run jobs on-demand. This commit adds a new `JobEmitter` interface
which is an EventEmitter which emits job events. The job system will
listen to these events and run the jobs as they come in.