build: Fix electron rebuild configuration
Since we no longer list our native dependencies as `dependencies`,
electron-rebuild won't rebuild them by default. Explicitly list
the native dependencies in the electron-rebuild configuration and force
a rebuild of the native dependencies when building the electron app in
order to make switching between running the app and the tests less
tedious.