feat: First _rough_ draft of HTML game integration
Add a new service which will be responsible for managing `BrowserWindow`
instances for HTML games. This service will be responsible for creating
new windows, managing their lifecycle, etc.

The `game:` protocol definition will need to be moved into the `pal`
module and can use further abstraction.

We will need to extend the database to store a launch path for HTML
games. Not all games use the (currently hardcoded) `index.html` file as
the entry point.

All in all this is a _very_ rough draft and will need to be refined.